feat(junit):  Implement screenshots via fixtures
