+++ playwright/src/test/java/com/microsoft/playwright/junit/TestScreenshotWithCustomOutputPath.java
@@ -3,7 +3,7 @@ package com.microsoft.playwright.junit;
 import com.microsoft.playwright.Page;
 import org.junit.jupiter.api.Test;
 
-import static com.microsoft.playwright.junit.FixtureTestUtils.getCustomOutputPath;
+import static com.microsoft.playwright.junit.TestTracesAndScreenshots.getCustomOutputPath;
 import static org.junit.jupiter.api.Assertions.fail;
 import static org.junit.jupiter.api.Assumptions.abort;
 
@@ -13,7 +13,7 @@ public class TestScreenshotWithCustomOutputPath {
   public static class TestOptions implements OptionsFactory {
     @Override
     public Options getOptions() {
-      return new Options().setScreenshot(Options.Screenshot.ON).setOutputDir(getCustomOutputPath().resolve("screenshotTest"));
+      return new Options().setScreenshot(Options.Screenshot.ON).setOutputDir(getCustomOutputPath());
     }
   }
